+++ src/firmware/lib/blinkscript/src/parser/ast/expressions/array_expression.hpp
@@ -11,15 +11,15 @@ namespace SyncBlink
     class ArrayExpression : public AstNode
     {
     public:
-        explicit ArrayExpression(std::vector<std::unique_ptr<const AstNode>>&& arrayContent, uint16_t line);
+        explicit ArrayExpression(std::vector<std::unique_ptr<const AstNode>>&& arrayContent, size_t line);
 
         void accept(AstVisitor& visitor) const;
-        uint16_t getLine() const;
+        size_t getLine() const;
         const std::vector<std::unique_ptr<const AstNode>>& getArrayContent() const;
 
     private:
         std::vector<std::unique_ptr<const AstNode>> _arrayContent;
-        uint16_t _line;
+        size_t _line;
     };
 }
