+++ src/firmware/lib/blinkscript/src/parser/ast/expressions/condition_expression.cpp
@@ -3,7 +3,7 @@
 namespace SyncBlink
 {
     ConditionExpression::ConditionExpression(std::unique_ptr<const AstNode> condition, std::unique_ptr<const AstNode> ifStatements,
-                                             std::unique_ptr<const AstNode> elseStatements, uint16_t line)
+                                             std::unique_ptr<const AstNode> elseStatements, size_t line)
         : _condition(std::move(condition)), _ifStatements(std::move(ifStatements)), _elseStatements(std::move(elseStatements)), _line(line)
     {
     }
@@ -28,7 +28,7 @@ namespace SyncBlink
         return _elseStatements;
     }
 
-    uint16_t ConditionExpression::getLine() const
+    size_t ConditionExpression::getLine() const
     {
         return _line;
     }
