+++ src/firmware/lib/blinkscript/src/parser/ast/expressions/condition_expression.hpp
@@ -9,19 +9,19 @@ namespace SyncBlink
     {
     public:
         ConditionExpression(std::unique_ptr<const AstNode> condition, std::unique_ptr<const AstNode> ifStatements,
-                            std::unique_ptr<const AstNode> elseStatements, uint16_t line);
+                            std::unique_ptr<const AstNode> elseStatements, size_t line);
 
         void accept(AstVisitor& visitor) const;
         const AstNode& getCondition() const;
         const AstNode& getIfBranch() const;
         const std::unique_ptr<const AstNode>& getElseBranch() const;
-        uint16_t getLine() const;
+        size_t getLine() const;
 
     private:
         std::unique_ptr<const AstNode> _condition;
         std::unique_ptr<const AstNode> _ifStatements;
         std::unique_ptr<const AstNode> _elseStatements;
-        uint16_t _line;
+        size_t _line;
     };
 }
