+++ src/firmware/lib/blinkscript/src/parser/ast/expressions/while_expression.hpp
@@ -8,17 +8,17 @@ namespace SyncBlink
     class WhileExpression : public AstNode
     {
     public:
-        WhileExpression(std::unique_ptr<const AstNode> condition, std::unique_ptr<const AstNode> loopBody, uint16_t line);
+        WhileExpression(std::unique_ptr<const AstNode> condition, std::unique_ptr<const AstNode> loopBody, size_t line);
 
         void accept(AstVisitor& visitor) const;
         const AstNode& getCondition() const;
         const AstNode& getLoopBody() const;
-        uint16_t getLine() const;
+        size_t getLine() const;
 
     private:
         std::unique_ptr<const AstNode> _condition;
         std::unique_ptr<const AstNode> _loopBody;
-        uint16_t _line;
+        size_t _line;
     };
 }
