+++ src/firmware/lib/blinkscript/src/program/model/objects/native_function_object.cpp
@@ -2,18 +2,20 @@
 
 namespace SyncBlink
 {
-    NativeFunObj::NativeFunObj(NativeFunc nativeFun, uint8_t arity) : _nativeFun(nativeFun), _arity(arity)
+    NativeFunObj::NativeFunObj(NativeFunc nativeFun, size_t arity) : _nativeFun(nativeFun), _arity(arity)
     {
     }
 
-    uint8_t NativeFunObj::getArity() const
+    size_t NativeFunObj::getArity() const
     {
         return _arity;
     }
+
     NativeFunc NativeFunObj::getFun() const
     {
         return _nativeFun;
     }
+    
     ObjectType NativeFunObj::getType() const
     {
         return ObjectType::NATIVEFUN;
